// src/app/web/api/tranUserSeachDivition.ts
export interface TranUserSeachDivition
{
    divId: number;
    divParentId?: any;
    orgId: number;
    orgName?: any;
    locId: number;
    locName?: any;
    locCode?: any;
    divisionCode: string;
    divisionName: string;
    divisionStatus: string;
    divisionType: string;
    partnerFlag: string;
    createdDate: string;
    createdBy: string;
    updatedDate: string;
    updatedBy: string;

}